Combine ReducedExperiment objects by columns

Combines ReducedExperiment objects by columns (samples). This method assumes that objects have identical features and components (i.e., factors or modules). If they are not, an error is returned.

So, this means that the feature-level slots should be equivalent, for example the assay rownames and the rownames of the factor loadings available in FactorisedExperiment objects. The component slots should also be equivalent, such as the column names of the reduced matrix or the column names of the aformentioned factor loadings matrix.

# S4 method for class 'FactorisedExperiment'
cbind(..., deparse.level = 1)

# S4 method for class 'ModularExperiment'
cbind(..., deparse.level = 1)

# S4 method for class 'ReducedExperiment'
cbind(..., deparse.level = 1)

Arguments

...

A series of ReducedExperiment objects to be combined. See cbind,SummarizedExperiment-method for further details.

deparse.level

Integer, see cbind.

Value

Returns a single ReducedExperiment object containing all of the columns in the objects passed to cbind.

Details

The SummarizedExperiment package includes separate methods for cbind (cbind,SummarizedExperiment-method) and (combineRows). The latter is supposed to be more flexible, permitting differences in the number and identity of the rows. For ReducedExperiment objects we only implement a single, less flexible, method that assumes the rows and components (i.e., factors or modules) are identical across objects. Attempting to apply combineRows to a ReducedExperiment object will result in the objects being treated as if they were SummarizedExperiments, and a single SummarizedExperiment object will be returned.

See also

Examples

# Create randomised containers with different numbers of samples
i <- 300 # Number of features
k <- 10 # Number of components (i.e., factors/modules)

# Same features and components, different samples (30 vs. 50 columns)
re_1 <- ReducedExperiment:::.createRandomisedReducedExperiment(i, 50, k)
re_2 <- ReducedExperiment:::.createRandomisedReducedExperiment(i, 30, k)

# Make a new object with 80 columns
cbind(re_1, re_2)
#> class: ReducedExperiment 
#> dim: 300 80 10 
#> metadata(0):
#> assays(1): normal
#> rownames(300): gene_1 gene_2 ... gene_299 gene_300
#> rowData names(0):
#> colnames(80): sample_1 sample_2 ... sample_29 sample_30
#> colData names(0):
#> 10 components

# We can apply combineRows to `ReducedExperiment` objects but the resulting
# object will be a `SummarizedExperiment`
combineRows(re_1, re_2)
#> class: SummarizedExperiment 
#> dim: 600 50 
#> metadata(0):
#> assays(1): normal
#> rownames(600): gene_1 gene_2 ... gene_299 gene_300
#> rowData names(0):
#> colnames(50): sample_1 sample_2 ... sample_49 sample_50
#> colData names(0):