Functional enrichment analyses for dimensionally-reduced data

Functional enrichment analyses for dimensionally-reduced data

# S4 method for FactorisedExperiment
runEnrich(
  object,
  method = "gsea",
  feature_id_col = "rownames",
  center_loadings = FALSE,
  abs_loadings = FALSE,
  loading_threshold = 0.5,
  proportional_threshold = 0.01,
  as_dataframe = FALSE,
  ...
)

# S4 method for ModularExperiment
runEnrich(
  object,
  method = "overrepresentation",
  feature_id_col = "rownames",
  as_dataframe = FALSE,
  ...
)

Arguments

object

FactorisedExperiment or ModularExperiment object.

method

The method to use for identifying enriched pathways. One of "overrepresentation" or "gsea". The overrepresentation method calls enricher whereas the gsea method calls GSEA. Note that GSEA is not available for modules.

feature_id_col

The column in rowData(object) that will be used as a feature ID. Setting this to "rownames" (default) instead uses rownames(object).

center_loadings

Factors only: If TRUE, loadings will be centered column-wise to have a mean of 0.

abs_loadings

Factors only: If TRUE, the absolute values of the loadings will be used for enrichment analysis. If FALSE, the signed loadings will be used for GSEA enrichment. Note that, regardless of the value of this term, the process used to select genes for overrepresentation analysis will be based on absolute loadings.

loading_threshold

Factors only: See getAlignedFeatures. Only relevant for overresentation analysis.

proportional_threshold

Factors only: See getAlignedFeatures. Only relevant for overrepresentation analysis.

as_dataframe

If TRUE, the results will be returned as a data.frame. Otherwise, the results will be returned as a list of objects created by either enricher, in the case of overrepresentation analysis, or GSEA, in the case of GSEA.

...

Additional arguments to be passed to GSEA (if method == "gsea") or enricher (if method == "overrepresentation").

Details

When running module analysis, the overrepresentation method identifies pathways that are overrepresented in each module.

For factor analysis, the overrepresentation method first identifies the genes most highly aligned with each factor (using getAlignedFeatures), then uses the resulting gene lists to perform overrepresentation analysis. The GSEA method instead uses the entire set of factor loadings, and identifies pathways that are overrepresented in the tails of this distribution.