Get correlation of features with module eigengenes

Provides a wrapper around signedKME. Provides a measure of module centrality/connectivity of each feature. Calculates correlation (Pearson's r) of each feature with the module eigengene (i.e., the column of reduced to which the feature belongs).

# S4 method for ModularExperiment
getCentrality(object, assay_name = "normal", feature_id_col = "rownames")

Arguments

object

A ModularExperiment object.

assay_name

The name of the assay to be used for calculation of module centrality.

feature_id_col

The column in rowData(object) that will be used as a feature ID. Setting this to "rownames" (default) instead uses rownames(object).

Value

Returns a data.frame with columns for feature, r (signed correlation with the eigengene), rsq (squared correlation with the eigengene), rank_r (feature rank based on r) and rank_rsq (feature rank based on rsq).

See also

Author

Jack Gisby

Examples

# Create ModularExperiment with random data (100 features, 50 samples,
# 10 modules)
me <- ReducedExperiment:::.createRandomisedModularExperiment(100, 50, 10)
me
#> class: ModularExperiment 
#> dim: 100 50 10 
#> metadata(0):
#> assays(1): normal
#> rownames(100): gene_1 gene_2 ... gene_99 gene_100
#> rowData names(0):
#> colnames(50): sample_1 sample_2 ... sample_49 sample_50
#> colData names(0):
#> 10 components

# Calculate centrality of each feature for the corresponding module
head(getCentrality(me))
#>        module  feature           r         rsq rank_r rank_rsq
#> 2    module_1 gene_100 -0.19565938 0.038282591      2        1
#> 1    module_1  gene_32  0.06879751 0.004733098      1        2
#> 99  module_10  gene_49 -0.18242115 0.033277474      4        1
#> 97  module_10  gene_17 -0.13471658 0.018148558      3        2
#> 98  module_10  gene_39  0.11412654 0.013024868      1        3
#> 100 module_10  gene_76  0.05991405 0.003589694      2        4