Apply dimensionality reduction using WGCNA

Performs Weighted gene correlation network analysis (WGCNA). Calls ReducedExperimentrun_wgcna to perform the analysis.

identify_modules(
  X,
  center_X = TRUE,
  scale_X = TRUE,
  assay_name = "normal",
  ...
)

Arguments

X

Either a SummarizedExperimentSummarizedExperiment object or a matrix containing data to be subject to WGCNA. X should have rows as features and columns as samples.

center_X

If TRUE, X is centered (i.e., features / rows are transformed to have a mean of 0) prior to WGCNA.

scale_X

If TRUE, X is scaled (i.e., features / rows are transformed to have a standard deviation of 1) before WGCNA.

assay_name

If X is a SummarizedExperimentSummarizedExperiment, then this should be the name of the assay to be subject to WGCNA.

...

Additional arguments to be passed to ReducedExperimentrun_wgcna.