ModularExperiment: A container for the results of module analysis

A container inheriting from ReducedExperiment, that contains one or more data matrices, to which module analysis has been applied to identify a reduced set of features.

ModularExperiment(
  reduced = new("matrix"),
  scale = TRUE,
  center = TRUE,
  loadings = numeric(),
  assignments = character(),
  dendrogram = NULL,
  threshold = NULL,
  ...
)

Arguments

reduced

A data matrix, produced by module analysis, with rows representing samples and columns representing module expression profiles. Typically, this matrix contains "eigengenes" produced by the Weighted Gene Correlation Network Analysis Approach, as is applied by identify_modules.

scale

Either a boolean, representing whether or not the original data has been scaled to unit variance, or a numeric vector indicating the standard deviations of the original features (as produced by scale.)

center

Either a boolean, representing whether or not the original data has been centered to have a mean of 0, or a numeric vector indicating the means of the original features (as produced by scale.)

loadings

A numeric vector representing the loadings used to generate module expression profiles. Typically, these values are obtained from the rotation matrix produced by prcomp, which is used to identify the first principal component of each module. The vector names represent features.

assignments

A vector of features, named according to the module to which the feature belongs.

dendrogram

Either NULL, or the dendrogram used to identify modules from the original data.

threshold

Either NULL, or a matrix produced by pickSoftThreshold indicating the parameters used for network construction.

...

Additional arguments to be passed to ReducedExperiment.