Plot a dendrogram stored in a ModularExperiment

Gets the dendrogram and plots it using plotDendroAndColors.

# S4 method for class 'ModularExperiment'
plotDendro(
  object,
  groupLabels = "Module colors",
  dendroLabels = FALSE,
  hang = 0.03,
  addGuide = TRUE,
  guideHang = 0.05,
  color_func = WGCNA::labels2colors,
  modules_are_colors = FALSE,
  ...
)

Arguments

object

ModularExperiment object.

groupLabels

Module label axis label. See plotDendroAndColors.

dendroLabels

If TRUE, shows feature names in the dendrogram. See plotDendroAndColors.

hang

The fraction of the plot height by which labels should hang below the rest of the plot. See plot.hclust.

addGuide

If TRUE, adds vertical guide lines to the dendrogram. See plotDendroAndColors.

guideHang

The fraction of the dendrogram's height to leave between the top end of the guide line and the dendrogram merge height. See plotDendroAndColors.

color_func

Function for converting module names to colors. Only used if modules_are_colors is FALSE

modules_are_colors

If TRUE, expects the module names to be colors. Else, assumes that module names are are numbers that can be converted into colours by color_func.

...

Additional arguments to be passed to plotDendroAndColors.