Plot module preservation statistics

Plot module preservation statistics

plot_module_preservation(
  module_preservation_results,
  show_random = TRUE,
  remove_module = NULL
)

Arguments

module_preservation_results

The output of module_preservation

show_random

If TRUE, shows the random module in the plots.

remove_module

The name of a module to be hidden from the plots.

Value

Two ggplot2 plot objects combined by patchwork. Plots the module preservation statistics generated by module_preservation.

Examples

# Get random matrices with rnorm, with 100 rows (features)
# and 20 columns (observations)
X_1 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")
X_2 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")

# Estimate 5 factors based on the data matrix
me_1 <- identify_modules(X_1, nc = 5, verbose = 0, powers = 5)
#>   Power SFT.R.sq slope truncated.R.sq mean.k. median.k. max.k.
#> 1     5     0.28  -102         0.0754   0.506     0.505  0.548
#> Warning: blockwiseModules: mergeCloseModules failed with the following error message:
#>      Error in mergeCloseModules(datExpr, colors[gsg$goodGenes], cutHeight = mergeCutHeight,  : 
#>   Error in moduleEigengenes(expr = exprData[[set]]$data, colors = setColors,  : 
#>   Color levels are empty. Possible reason: the only color is grey and grey module is excluded from the calculation.
#> 
#>  
#> --> returning unmerged colors.
me_2 <- identify_modules(X_2, nc = 5, verbose = 0, powers = 5)
#>   Power SFT.R.sq slope truncated.R.sq mean.k. median.k. max.k.
#> 1     5    0.177   -58        -0.0409   0.504     0.501  0.546
#>      mergeCloseModules: less than two proper modules.
#>       ..color levels are blue, grey
#>       ..there is nothing to merge.

# Test module preservation
mp <- module_preservation(me_1, me_2, verbose = 10, nPermutations = 10)
#>   ..checking data for excessive amounts of missing data..
#>      Flagging genes and samples with too many missing values...
#>       ..step 1
#>      Flagging genes and samples with too many missing values...
#>       ..step 1
#>   ..unassigned 'module' name: module_0 
#>   ..all network sample 'module' name: random
#>   ..calculating observed preservation values
#>      ..working on reference network reference
#>        ..working on test network test
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>   ..calculating permutation Z scores
#>  ..Working with set 1 as reference set
#>  *+*+*+*+* Reference set contains a single module.
#>   A permutation analysis is not meaningful for a single module; skipping.
#> Error in WGCNA::modulePreservation(multi_data, list(reference = stats::setNames(names(module_assignments),     module_assignments)), dataIsExpr = TRUE, networkType = networkType,     corFnc = corFnc, goldName = goldName, greyName = greyName,     savePermutedStatistics = savePermutedStatistics, ...): object 'connSummaryInd' not found

# Plot the results
plot_module_preservation(mp)
#> Error in eval(expr, envir, enclos): object 'mp' not found