Assess soft thresholding power for WGCNA

A wrapper around pickSoftThreshold allowing assessment and automatic selection of soft-thresholding power. Extends the function to accept a SummarizedExperiment as input and additionally considers mean connectivity when selecting the soft-thresholding power to recommend.

assess_soft_threshold(
  X,
  assay_name = "normal",
  powerVector = 1:30,
  RsquaredCut = 0.85,
  max_mean_connectivity = 100,
  corType = "pearson",
  networkType = "signed",
  maxBlockSize = 30000,
  verbose = 0,
  ...
)

Arguments

X

Either a SummarizedExperiment object or a matrix containing data to be subject to WGCNA. X should have rows as features and columns as samples.

assay_name

If X is a SummarizedExperiment, then this should be the name of the assay to be subject to WGCNA.

powerVector

A vector of integers indicating soft-thresholding powers to be assessed. See pickSoftThreshold.

RsquaredCut

The minimum scale free topology fitting index required. Used to select the soft-thresholding power. See pickSoftThreshold.

max_mean_connectivity

The maximal mean connectivity required. Used to select the soft-thresholding power.

corType

The type of correlation to be used to generate a correlation matrix during network formation. One of "pearson" (cor) and "bicor" (bicor).

networkType

The type of network to be generated, either "unsigned", "signed" or "signed hybrid". See pickSoftThreshold for more details.

maxBlockSize

The chunk size (in terms of the number of features/genes) to process the data. See pickSoftThreshold for more details. The default (30000) should process standard transcriptomic datasets in a single chunk. Results may differ if the number of features exceeds the chunk size. Lower values of this parameter will use less memory to calculate networks.

verbose

The verbosity, passed to pickSoftThreshold.

...

Additional arguments to be passed to pickSoftThreshold.

Value

Returns a data.frame generated by pickSoftThreshold with scale free topology fitting indices and connectivity statistics. Additionally contains a column, "estimated_power", indicating the recommended power to use (see details). We suggest manually considering suitability of the soft-thresholding power rather than solely relying on this automated approach.

Details

The pickSoftThreshold function estimates the power by selecting the lowest value that with a minimum scale free topology fitting index exceeding RsquaredCut. The assess_soft_threshold function mirrors this behaviour when max_mean_connectivity is NULL. When max_mean_connectivity is specified, however, we additionally ensure that the selected power does not exceed this connectivity threshold.

See also

Author

Jack Gisby