Runs linear models for components and sample-level data

Runs either standard linear or linear mixed models between reduced components (e.g., factors or modules) and sample-level information.

associate_components(
  re,
  formula,
  method = "lm",
  scale_reduced = TRUE,
  center_reduced = TRUE,
  type = "II",
  adj_method = "BH",
  ...
)

Arguments

re

An object inheriting from ReducedExperiment.

formula

The model formula to apply. Only the right hand side of the model need be specified (e.g., "~ x + y"). The left hand side represents the components themselves. The variables in this formula should be present in the colData of re.

method

If "lm", then the lm function is used to run linear models (in tandem with Anova for running anovas on the model terms). If "lmer", then linear mixed models are run through lmer.

scale_reduced

If TRUE, the reduced data are scaled (to have a standard deviation of 1) before modelling.

center_reduced

If TRUE, the reduced data are centered (to have a mean of 0) before modelling.

type

The type of anova to be applied to the terms of the linear model.

adj_method

The method for adjusting for multiple testing. Passed to the p.adjust method parameter.

...

Additional arguments passed to lmer, given that method is set to "lmer".

Value

Returns a list with the entry "models" including a list of the model objects, "anovas" containing the output of anova-based testing, and "summaries" containing the results of running summary on the models.