Functional enrichment analyses for dimensionally-reduced data

Functional enrichment analyses for dimensionally-reduced data

# S4 method for class 'FactorisedExperiment'
runEnrich(
  object,
  method = "gsea",
  feature_id_col = "rownames",
  center_loadings = FALSE,
  abs_loadings = FALSE,
  loading_threshold = 0.5,
  proportional_threshold = 0.01,
  as_dataframe = FALSE,
  ...
)

# S4 method for class 'ModularExperiment'
runEnrich(
  object,
  method = "overrepresentation",
  feature_id_col = "rownames",
  as_dataframe = FALSE,
  ...
)

Arguments

object

FactorisedExperiment or ModularExperiment object.

method

The method to use for identifying enriched pathways. One of "overrepresentation" or "gsea". The overrepresentation method calls enricher whereas the gsea method calls GSEA. Note that GSEA is not available for modules.

feature_id_col

The column in rowData(object) that will be used as a feature ID. Setting this to "rownames" (default) instead uses rownames(object).

center_loadings

Factors only: If TRUE, loadings will be centered column-wise to have a mean of 0.

abs_loadings

Factors only: If TRUE, the absolute values of the loadings will be used for enrichment analysis. If FALSE, the signed loadings will be used for GSEA enrichment. Note that, regardless of the value of this term, the process used to select genes for overrepresentation analysis will be based on absolute loadings.

loading_threshold

Factors only: See getAlignedFeatures. Only relevant for overresentation analysis.

proportional_threshold

Factors only: See getAlignedFeatures. Only relevant for overrepresentation analysis.

as_dataframe

If TRUE, the results will be returned as a data.frame. Otherwise, the results will be returned as a list of objects created by either enricher, in the case of overrepresentation analysis, or GSEA, in the case of GSEA.

...

Additional arguments to be passed to GSEA (if method == "gsea") or enricher (if method == "overrepresentation").

Value

If as_dataframe is TRUE, the results will be returned as a data.frame. Otherwise, the results will be returned as a list of objects created by either enricher, in the case of overrepresentation analysis, or GSEA, in the case of GSEA.

Details

When running module analysis, the overrepresentation method identifies pathways that are overrepresented in each module.

For factor analysis, the overrepresentation method first identifies the genes most highly aligned with each factor (using getAlignedFeatures), then uses the resulting gene lists to perform overrepresentation analysis. The GSEA method instead uses the entire set of factor loadings, and identifies pathways that are overrepresented in the tails of this distribution.

Author

Jack Gisby

Examples

set.seed(2)
airway <- ReducedExperiment:::.get_airway_data(n_features = 2000)
airway_fe <- estimate_factors(
    airway,
    nc = 2,
    use_stability = FALSE,
    method = "imax"
)

# Run overrepresentation analysis
overrep_res <- runEnrich(
    airway_fe,
    method = "overrepresentation",
    feature_id_col = "rownames",
    as_dataframe = TRUE,
    p_cutoff = 0.1,
    universe = rownames(airway_fe)
)

head(overrep_res)
#>                                                                           ID
#> factor_1.NABA_MATRISOME_ASSOCIATED                 NABA_MATRISOME_ASSOCIATED
#> factor_1.NABA_MATRISOME                                       NABA_MATRISOME
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                 WP_MAPK_SIGNALING_PATHWAY
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY             KEGG_MAPK_SIGNALING_PATHWAY
#> factor_2.REACTOME_NEURONAL_SYSTEM                   REACTOME_NEURONAL_SYSTEM
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA WP_MALIGNANT_PLEURAL_MESOTHELIOMA
#>                                                                  Description
#> factor_1.NABA_MATRISOME_ASSOCIATED                 NABA_MATRISOME_ASSOCIATED
#> factor_1.NABA_MATRISOME                                       NABA_MATRISOME
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                 WP_MAPK_SIGNALING_PATHWAY
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY             KEGG_MAPK_SIGNALING_PATHWAY
#> factor_2.REACTOME_NEURONAL_SYSTEM                   REACTOME_NEURONAL_SYSTEM
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA WP_MALIGNANT_PLEURAL_MESOTHELIOMA
#>                                            GeneRatio BgRatio      pvalue
#> factor_1.NABA_MATRISOME_ASSOCIATED              4/11 52/1010 0.001579609
#> factor_1.NABA_MATRISOME                         4/11 69/1010 0.004572905
#> factor_2.WP_MAPK_SIGNALING_PATHWAY              3/13 22/1010 0.002231104
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY            3/13 24/1010 0.002888557
#> factor_2.REACTOME_NEURONAL_SYSTEM               3/13 31/1010 0.006085876
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA      3/13 39/1010 0.011649264
#>                                               p.adjust qvalue
#> factor_1.NABA_MATRISOME_ASSOCIATED         0.003159217     NA
#> factor_1.NABA_MATRISOME                    0.004572905     NA
#> factor_2.WP_MAPK_SIGNALING_PATHWAY         0.008665670     NA
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY       0.008665670     NA
#> factor_2.REACTOME_NEURONAL_SYSTEM          0.012171751     NA
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA 0.017473897     NA
#>                                                                                                     geneID
#> factor_1.NABA_MATRISOME_ASSOCIATED         ENSG00000172156/ENSG00000123243/ENSG00000145423/ENSG00000126838
#> factor_1.NABA_MATRISOME                    ENSG00000172156/ENSG00000123243/ENSG00000145423/ENSG00000126838
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                         ENSG00000066468/ENSG00000165995/ENSG00000173110
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY                       ENSG00000066468/ENSG00000165995/ENSG00000173110
#> factor_2.REACTOME_NEURONAL_SYSTEM                          ENSG00000099337/ENSG00000165995/ENSG00000135821
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA                 ENSG00000145423/ENSG00000040731/ENSG00000066468
#>                                            Count adj_method             method
#> factor_1.NABA_MATRISOME_ASSOCIATED             4         BH overrepresentation
#> factor_1.NABA_MATRISOME                        4         BH overrepresentation
#> factor_2.WP_MAPK_SIGNALING_PATHWAY             3         BH overrepresentation
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY           3         BH overrepresentation
#> factor_2.REACTOME_NEURONAL_SYSTEM              3         BH overrepresentation
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA     3         BH overrepresentation
#>                                            component loading_threshold
#> factor_1.NABA_MATRISOME_ASSOCIATED          factor_1               0.5
#> factor_1.NABA_MATRISOME                     factor_1               0.5
#> factor_2.WP_MAPK_SIGNALING_PATHWAY          factor_2               0.5
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY        factor_2               0.5
#> factor_2.REACTOME_NEURONAL_SYSTEM           factor_2               0.5
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA  factor_2               0.5
#>                                            proportional_threshold
#> factor_1.NABA_MATRISOME_ASSOCIATED                           0.01
#> factor_1.NABA_MATRISOME                                      0.01
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                           0.01
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY                         0.01
#> factor_2.REACTOME_NEURONAL_SYSTEM                            0.01
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA                   0.01
#>                                            loadings_centered loadings_scaled
#> factor_1.NABA_MATRISOME_ASSOCIATED                     FALSE            TRUE
#> factor_1.NABA_MATRISOME                                FALSE            TRUE
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                     FALSE            TRUE
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY                   FALSE            TRUE
#> factor_2.REACTOME_NEURONAL_SYSTEM                      FALSE            TRUE
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA             FALSE            TRUE
#>                                            abs_loadings
#> factor_1.NABA_MATRISOME_ASSOCIATED                FALSE
#> factor_1.NABA_MATRISOME                           FALSE
#> factor_2.WP_MAPK_SIGNALING_PATHWAY                FALSE
#> factor_2.KEGG_MAPK_SIGNALING_PATHWAY              FALSE
#> factor_2.REACTOME_NEURONAL_SYSTEM                 FALSE
#> factor_2.WP_MALIGNANT_PLEURAL_MESOTHELIOMA        FALSE

# Run gene set enrichment approach
gsea_res <- runEnrich(
    airway_fe,
    method = "gsea",
    feature_id_col = "rownames",
    as_dataframe = TRUE,
    p_cutoff = 0.1
)
#> using 'fgsea' for GSEA analysis, please cite Korotkevich et al (2019).
#> preparing geneSet collections...
#> GSEA analysis...
#> leading edge analysis...
#> done...
#> using 'fgsea' for GSEA analysis, please cite Korotkevich et al (2019).
#> preparing geneSet collections...
#> GSEA analysis...
#> leading edge analysis...
#> done...

head(gsea_res)
#>                                                          ID
#> factor_2 WP_OSTEOBLAST_DIFFERENTIATION_AND_RELATED_DISEASES
#>                                                 Description setSize
#> factor_2 WP_OSTEOBLAST_DIFFERENTIATION_AND_RELATED_DISEASES      12
#>          enrichmentScore     NES      pvalue   p.adjust qvalue rank
#> factor_2       0.8344876 1.95485 0.000136704 0.03499623     NA  175
#>                           leading_edge
#> factor_2 tags=58%, list=9%, signal=54%
#>                                                                                                          core_enrichment
#> factor_2 ENSG00000066468/ENSG00000002745/ENSG00000133056/ENSG00000170365/ENSG00000109339/ENSG00000090932/ENSG00000074047
#>          adj_method method component loadings_centered loadings_scaled
#> factor_2         BH   gsea  factor_2             FALSE            TRUE
#>          abs_loadings
#> factor_2        FALSE