Get correlation of features with module eigengenes

Provides a wrapper around signedKME. Provides a measure of module centrality/connectivity of each feature. Essentially just calculates correlation (Pearson's r) of each feature with the module eigengene, i.e., the column of reduced to which the feature belongs.

# S4 method for class 'ModularExperiment'
getCentrality(object, assay_name = "normal", feature_id_col = "rownames")

Arguments

object

ModularExperiment object.

assay_name

The name of the assay to be used for calculation of module centrality.

feature_id_col

The column in rowData(object) that will be used as a feature ID. Setting this to "rownames" (default) instead uses rownames(object).

Value

Returns a data.frame with columns for feature, r (signed correlation with the eigengene), rsq (squared correlation with the eigengene), rank_r (feature rank based on r) and rank_rsq (feature rank based on rsq).

See also

Examples

# Create ModularExperiment with random data (100 features, 50 samples,
# 10 modules)
me <- ReducedExperiment:::.createRandomisedModularExperiment(100, 50, 10)
me
#> class: ModularExperiment 
#> dim: 100 50 10 
#> metadata(0):
#> assays(1): normal
#> rownames(100): gene_1 gene_2 ... gene_99 gene_100
#> rowData names(0):
#> colnames(50): sample_1 sample_2 ... sample_49 sample_50
#> colData names(0):
#> 10 components

# Calculate centrality of each feature for the corresponding module
getCentrality(me)
#>        module  feature             r          rsq rank_r rank_rsq
#> 2    module_1 gene_100 -0.1956593759 3.828259e-02      2        1
#> 1    module_1  gene_32  0.0687975143 4.733098e-03      1        2
#> 99  module_10  gene_49 -0.1824211450 3.327747e-02      4        1
#> 97  module_10  gene_17 -0.1347165842 1.814856e-02      3        2
#> 98  module_10  gene_39  0.1141265423 1.302487e-02      1        3
#> 100 module_10  gene_76  0.0599140539 3.589694e-03      2        4
#> 7    module_2  gene_50 -0.2646875572 7.005950e-02     10        1
#> 8    module_2  gene_52  0.1860125579 3.460067e-02      1        2
#> 9    module_2  gene_53 -0.1538561005 2.367170e-02      9        3
#> 3    module_2  gene_14 -0.1382606935 1.911602e-02      8        4
#> 5    module_2  gene_44  0.1380557006 1.905938e-02      2        5
#> 12   module_2  gene_99 -0.0956125374 9.141757e-03      7        6
#> 11   module_2  gene_97  0.0845472414 7.148236e-03      3        7
#> 4    module_2  gene_28 -0.0490892394 2.409753e-03      6        8
#> 10   module_2  gene_74 -0.0211072852 4.455175e-04      5        9
#> 6    module_2  gene_46 -0.0078914369 6.227478e-05      4       10
#> 15   module_3  gene_31 -0.4338466216 1.882229e-01      7        1
#> 13   module_3   gene_8 -0.1866051944 3.482150e-02      6        2
#> 18   module_3  gene_61 -0.1416652992 2.006906e-02      5        3
#> 17   module_3  gene_40 -0.1296828231 1.681763e-02      4        4
#> 14   module_3   gene_9 -0.1192568628 1.422220e-02      3        5
#> 19   module_3  gene_95  0.1114730608 1.242624e-02      1        6
#> 16   module_3  gene_34  0.0616305709 3.798327e-03      2        7
#> 26   module_4  gene_94 -0.3161317694 9.993930e-02      7        1
#> 24   module_4  gene_83 -0.1787338732 3.194580e-02      6        2
#> 22   module_4  gene_65  0.1659571919 2.754179e-02      1        3
#> 25   module_4  gene_89 -0.1195606116 1.429474e-02      5        4
#> 23   module_4  gene_69 -0.0958846328 9.193863e-03      4        5
#> 21   module_4  gene_25 -0.0429057798 1.840906e-03      3        6
#> 20   module_4   gene_1 -0.0308992605 9.547643e-04      2        7
#> 45   module_5  gene_87 -0.3303158936 1.091086e-01     20        1
#> 39   module_5  gene_63  0.2885297428 8.324941e-02      1        2
#> 33   module_5  gene_36  0.2608197076 6.802692e-02      2        3
#> 41   module_5  gene_71 -0.1834494774 3.365371e-02     19        4
#> 42   module_5  gene_81 -0.1714642362 2.939998e-02     18        5
#> 29   module_5  gene_15  0.1613789601 2.604317e-02      3        6
#> 44   module_5  gene_86  0.1580829746 2.499023e-02      4        7
#> 27   module_5   gene_5 -0.1544368387 2.385074e-02     17        8
#> 36   module_5  gene_54  0.1260912894 1.589901e-02      5        9
#> 40   module_5  gene_67 -0.1155604857 1.335423e-02     16       10
#> 37   module_5  gene_57 -0.1154779814 1.333516e-02     15       11
#> 46   module_5  gene_88  0.1143930067 1.308576e-02      6       12
#> 30   module_5  gene_24  0.0965177977 9.315685e-03      7       13
#> 28   module_5  gene_10 -0.0889537743 7.912774e-03     14       14
#> 34   module_5  gene_38 -0.0699969651 4.899575e-03     13       15
#> 31   module_5  gene_29 -0.0698758049 4.882628e-03     12       16
#> 35   module_5  gene_47 -0.0641284428 4.112457e-03     11       17
#> 43   module_5  gene_84 -0.0615689409 3.790734e-03     10       18
#> 32   module_5  gene_30 -0.0322973431 1.043118e-03      9       19
#> 38   module_5  gene_59 -0.0127580991 1.627691e-04      8       20
#> 56   module_6  gene_78 -0.3029850405 9.179993e-02     11        1
#> 49   module_6  gene_27  0.2373321604 5.632655e-02      1        2
#> 48   module_6  gene_13  0.1979071065 3.916722e-02      2        3
#> 52   module_6  gene_41  0.1309048460 1.713608e-02      3        4
#> 47   module_6  gene_12 -0.1229688281 1.512133e-02     10        5
#> 53   module_6  gene_56  0.1029426700 1.059719e-02      4        6
#> 54   module_6  gene_64 -0.0927109551 8.595321e-03      9        7
#> 57   module_6  gene_90 -0.0828222952 6.859533e-03      8        8
#> 51   module_6  gene_37 -0.0594682782 3.536476e-03      7        9
#> 55   module_6  gene_72 -0.0406095405 1.649135e-03      6       10
#> 50   module_6  gene_33 -0.0002471732 6.109459e-08      5       11
#> 61   module_7  gene_20  0.2318215723 5.374124e-02      1        1
#> 68   module_7  gene_85  0.2145396461 4.602726e-02      2        2
#> 59   module_7  gene_16  0.1988855048 3.955544e-02      3        3
#> 65   module_7  gene_60 -0.1870384359 3.498338e-02     13        4
#> 63   module_7  gene_43  0.1538040390 2.365568e-02      4        5
#> 70   module_7  gene_92  0.0613823514 3.767793e-03      5        6
#> 69   module_7  gene_91  0.0581690444 3.383638e-03      6        7
#> 66   module_7  gene_73  0.0492811675 2.428633e-03      7        8
#> 60   module_7  gene_18 -0.0266543507 7.104544e-04     12        9
#> 62   module_7  gene_35 -0.0217915278 4.748707e-04     11       10
#> 58   module_7   gene_6  0.0216042986 4.667457e-04      8       11
#> 67   module_7  gene_82 -0.0153631166 2.360254e-04     10       12
#> 64   module_7  gene_55  0.0098437549 9.689951e-05      9       13
#> 72   module_8   gene_3 -0.3784265606 1.432067e-01     19        1
#> 81   module_8  gene_45  0.3020578800 9.123896e-02      1        2
#> 74   module_8   gene_7  0.2212460970 4.894984e-02      2        3
#> 88   module_8  gene_93 -0.1882182642 3.542611e-02     18        4
#> 75   module_8  gene_11  0.1731297942 2.997393e-02      3        5
#> 86   module_8  gene_77 -0.1592721445 2.536762e-02     17        6
#> 82   module_8  gene_48  0.1555463286 2.419466e-02      4        7
#> 73   module_8   gene_4 -0.1438426870 2.069072e-02     16        8
#> 87   module_8  gene_80 -0.1341810212 1.800455e-02     15        9
#> 71   module_8   gene_2 -0.1094165129 1.197197e-02     14       10
#> 84   module_8  gene_68  0.1068737768 1.142200e-02      5       11
#> 77   module_8  gene_21 -0.0881035411 7.762234e-03     13       12
#> 85   module_8  gene_75 -0.0854918558 7.308857e-03     12       13
#> 76   module_8  gene_19  0.0640999985 4.108810e-03      6       14
#> 89   module_8  gene_96 -0.0616997101 3.806854e-03     11       15
#> 79   module_8  gene_23 -0.0573052766 3.283895e-03     10       16
#> 83   module_8  gene_62  0.0328761406 1.080841e-03      7       17
#> 78   module_8  gene_22  0.0214937868 4.619829e-04      8       18
#> 80   module_8  gene_42 -0.0095766769 9.171274e-05      9       19
#> 91   module_9  gene_51  0.1802270770 3.248180e-02      1        1
#> 95   module_9  gene_79  0.0813096076 6.611252e-03      2        2
#> 94   module_9  gene_70 -0.0600540181 3.606485e-03      7        3
#> 96   module_9  gene_98  0.0260239601 6.772465e-04      3        4
#> 90   module_9  gene_26 -0.0139982754 1.959517e-04      6        5
#> 92   module_9  gene_58  0.0098179113 9.639138e-05      4        6
#> 93   module_9  gene_66 -0.0080589337 6.494641e-05      5        7