Get and set module feature assignments

Retrieves a vector of features (usually genes) named by the modules they belong to. Assignment can be used to modify all or part of the vector.

Usage

# S4 method for class 'ModularExperiment'
assignments(object, as_list = FALSE)

# S4 method for class 'ModularExperiment'
assignments(object) <- value

Arguments

object

ModularExperiment object.

as_list

If TRUE, the results are returned as a list, with an entry for each module containing a list of features.

value

New value to replace existing assignments.

Value

A vector with values representing features and names representing feature assignments (i.e., modules).

Author

Jack Gisby

Examples

# Create ModularExperiment with random data (100 features, 50 samples,
# 10 modules)
me <- ReducedExperiment:::.createRandomisedModularExperiment(100, 50, 10)
me
#> class: ModularExperiment 
#> dim: 100 50 10 
#> metadata(0):
#> assays(1): normal
#> rownames(100): gene_1 gene_2 ... gene_99 gene_100
#> rowData names(0):
#> colnames(50): sample_1 sample_2 ... sample_49 sample_50
#> colData names(0):
#> 10 components

# Assignment of features to groups/modules
assignments(me)
#>   module_6   module_3   module_6   module_5   module_1   module_7   module_8 
#>   "gene_1"   "gene_2"   "gene_3"   "gene_4"   "gene_5"   "gene_6"   "gene_7" 
#>   module_2   module_5   module_6  module_10   module_7   module_5   module_1 
#>   "gene_8"   "gene_9"  "gene_10"  "gene_11"  "gene_12"  "gene_13"  "gene_14" 
#>   module_1   module_6   module_9   module_8   module_2   module_5   module_8 
#>  "gene_15"  "gene_16"  "gene_17"  "gene_18"  "gene_19"  "gene_20"  "gene_21" 
#>   module_9   module_7   module_3   module_4   module_8  module_10   module_5 
#>  "gene_22"  "gene_23"  "gene_24"  "gene_25"  "gene_26"  "gene_27"  "gene_28" 
#>   module_4   module_7   module_7   module_7   module_7   module_6   module_3 
#>  "gene_29"  "gene_30"  "gene_31"  "gene_32"  "gene_33"  "gene_34"  "gene_35" 
#>  module_10   module_6  module_10   module_6   module_2   module_4   module_9 
#>  "gene_36"  "gene_37"  "gene_38"  "gene_39"  "gene_40"  "gene_41"  "gene_42" 
#>   module_2   module_5   module_9   module_6   module_6   module_9   module_6 
#>  "gene_43"  "gene_44"  "gene_45"  "gene_46"  "gene_47"  "gene_48"  "gene_49" 
#>   module_7   module_5   module_3   module_7   module_1   module_6   module_8 
#>  "gene_50"  "gene_51"  "gene_52"  "gene_53"  "gene_54"  "gene_55"  "gene_56" 
#>   module_3   module_9   module_2   module_3   module_4   module_2   module_8 
#>  "gene_57"  "gene_58"  "gene_59"  "gene_60"  "gene_61"  "gene_62"  "gene_63" 
#>   module_8   module_5   module_2   module_1   module_2   module_7   module_7 
#>  "gene_64"  "gene_65"  "gene_66"  "gene_67"  "gene_68"  "gene_69"  "gene_70" 
#>   module_6   module_9   module_5   module_8   module_6   module_2  module_10 
#>  "gene_71"  "gene_72"  "gene_73"  "gene_74"  "gene_75"  "gene_76"  "gene_77" 
#>   module_5   module_3   module_4   module_9   module_4   module_3   module_9 
#>  "gene_78"  "gene_79"  "gene_80"  "gene_81"  "gene_82"  "gene_83"  "gene_84" 
#>   module_6   module_2   module_4   module_7   module_3   module_6   module_9 
#>  "gene_85"  "gene_86"  "gene_87"  "gene_88"  "gene_89"  "gene_90"  "gene_91" 
#>   module_7   module_9  module_10   module_5   module_9   module_9   module_4 
#>  "gene_92"  "gene_93"  "gene_94"  "gene_95"  "gene_96"  "gene_97"  "gene_98" 
#>   module_6   module_7 
#>  "gene_99" "gene_100" 

# We can reassign a feature to a new module if we like:
names(assignments(me))[6] <- "new_module"
assignments(me)[1:10]
#>   module_6   module_3   module_6   module_5   module_1 new_module   module_8 
#>   "gene_1"   "gene_2"   "gene_3"   "gene_4"   "gene_5"   "gene_6"   "gene_7" 
#>   module_2   module_5   module_6 
#>   "gene_8"   "gene_9"  "gene_10" 

# We shouldn't, however, attempt to change the feature names here:
# assignments(me)[5] <- "modified_gene_name"

# Instead, we should change the object's feature names as so:
featureNames(me)[5] <- "modified_gene_name"
assignments(me)[1:10]
#>             module_6             module_3             module_6 
#>             "gene_1"             "gene_2"             "gene_3" 
#>             module_5             module_1           new_module 
#>             "gene_4" "modified_gene_name"             "gene_6" 
#>             module_8             module_2             module_5 
#>             "gene_7"             "gene_8"             "gene_9" 
#>             module_6 
#>            "gene_10"