Project new data using pre-defined factors

Uses a projection approach to calculate factors in new data. Functions in a similar fashion to the predict method of prcomp. The transposed newdata are multiplied by the original loadings matrix.

# S4 method for FactorisedExperiment,matrix
projectData(
  object,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL
)

# S4 method for FactorisedExperiment,data.frame
projectData(
  object,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL
)

# S4 method for FactorisedExperiment,SummarizedExperiment
projectData(
  object,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL,
  assay_name = "normal"
)

# S4 method for FactorisedExperiment
predict(object, newdata, ...)

Arguments

object

A FactorisedExperiment object. The loadings slot of this class will be used for projection. Additionally, by default, the scale and center slots are used to apply the original transformation to the new data.

newdata

New data for projection. Must be a data.frame or matrix with features as rows and samples as columns, or a SummarizedExperiment object. Assumes that the rows of newdata match those of the FactorisedExperiment object.

scale_reduced

Whether or not the reduced data should be scaled after calculation.

scale_newdata

Controls whether the newdata are scaled. If NULL, performs scaling based on the FactorisedExperiment object's scale slot. The value of this argument will be passed to the scale argument of scale.

center_newdata

Controls whether the newdata are centered If NULL, performs centering based on the FactorisedExperiment object's center slot. The value of this argument will be passed to the center argument of scale.

assay_name

If a SummarizedExperiment object is passed as new data, this argument indicates which assay should be used for projection.

...

Additional arguments to be passed to projectData.

Value

Calculates a matrix with samples as rows and factors as columns. If newdata was a matrix or data.frame, this will be returned as a matrix. If a SummarizedExperiment object was passed instead, then a If a FactorisedExperiment

object will be created containing this matrix in its reduced slot.

See also

Examples

# Get two random matrices with rnorm
# 1: 100 rows (features) and 20 columns (observations)
X_1 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")

# Both matrices must have the same features, but they may have different obs
# 2: 100 rows (features) and 30 columns (observations)
X_2 <- ReducedExperiment:::.makeRandomData(100, 30, "feature", "obs")

# Estimate 5 factors based on the data matrix
fe_1 <- estimate_factors(X_1, nc = 5)
fe_1
#> class: FactorisedExperiment 
#> dim: 100 20 5 
#> metadata(0):
#> assays(2): normal transformed
#> rownames(100): feature_1 feature_2 ... feature_99 feature_100
#> rowData names(0):
#> colnames(20): obs_1 obs_2 ... obs_19 obs_20
#> colData names(0):
#> 5 components

# Project the fe_1 factors for the samples in X_2
fe_2 <- projectData(fe_1, X_2)
fe_2
#>           factor_1    factor_2     factor_3    factor_4    factor_5
#> obs_1  -0.05324222 -0.31555187  0.699147131 -1.00941522 -1.76451105
#> obs_2   0.83576421 -0.69547229  0.159117645  0.55592813  0.30579631
#> obs_3   1.13118474  2.41778204 -0.054443515  0.46792177  1.40456566
#> obs_4  -0.31856790 -1.66543082  0.521999799  1.14918808 -0.82430197
#> obs_5   0.66460005  0.22443362 -1.083005574 -0.75459052 -0.22691631
#> obs_6  -2.18913847  1.14329931  0.101001100  1.14115945  0.29373794
#> obs_7   1.17048007  0.93667782 -0.782244989  2.97485618  0.92144494
#> obs_8  -0.49172483 -0.32165945 -0.345583195 -0.07541566  0.17915930
#> obs_9  -0.39023242 -1.31983028  1.178914559  0.64781515  0.57460051
#> obs_10  0.52843885 -0.13879981 -0.358421693 -0.80183711  0.44004667
#> obs_11 -0.23938349  0.30869202 -0.592972191 -0.90805433  0.08694723
#> obs_12 -0.95624200 -0.79517663  0.746796527 -0.06278988  1.86727478
#> obs_13 -0.57119988 -0.39182304  0.497392172 -0.43963582 -1.20086721
#> obs_14  0.95121489  1.19588692  1.581185538  0.25746778  1.34191927
#> obs_15  0.42123089 -0.89537628  0.369596116 -0.42094557 -0.41452911
#> obs_16 -1.21850084 -1.30112706 -1.769879222  1.02781402  0.59105494
#> obs_17  0.43635793  0.56858368  2.332463200 -0.76192308 -1.46833307
#> obs_18  2.53467614  0.88763358 -1.736375784  1.05573568 -1.59956378
#> obs_19 -1.16598123  1.86534142  0.078075699  0.39150406  0.20379232
#> obs_20 -0.27438157  0.10560683  0.386156695 -1.25618312 -0.01167785
#> obs_21  0.12689653  1.35690758  0.006238685  0.29783938  0.95235792
#> obs_22 -1.57548497 -0.82673684  0.198194092 -1.15309439  2.13128538
#> obs_23 -1.04470054 -0.49144696  0.551479779 -0.08071957  0.25747883
#> obs_24  1.09129568 -0.06854656 -1.091235135 -0.84093237 -1.15724194
#> obs_25 -0.69643016 -1.10121190  0.099834381 -0.03897311 -0.11737642
#> obs_26  0.39220778 -0.70271761  0.770447069  0.99665522 -0.02672246
#> obs_27 -0.21737874  0.32500811 -1.647156912 -0.58221869 -0.70366027
#> obs_28  1.47714349 -1.10705881  0.850041265 -0.02025558 -0.12908276
#> obs_29 -0.25834033  0.11835403 -1.900761344 -2.21730815 -1.39182772
#> obs_30 -0.10056164  0.68375928  0.233998102  0.46040725 -0.51485007
#> attr(,"scaled:center")
#>  factor_1  factor_2  factor_3  factor_4  factor_5 
#>  1.048171 -1.850275 -2.801580  1.515847  5.759392 
#> attr(,"scaled:scale")
#>  factor_1  factor_2  factor_3  factor_4  factor_5 
#>  7.869577 11.105263 10.093258  9.995234 11.703628