Run WGCNA for a data matrix

Runs WGCNA. Largely a wrapper for the blockwiseModules function that reformats data into a format convenient for creating a ModularExperiment object and changes module names from colours to numbers by default.

Usage

run_wgcna(
  X,
  power,
  corType = "pearson",
  networkType = "signed",
  module_labels = "numbers",
  maxBlockSize = 30000,
  verbose = 0,
  standardise_reduced = TRUE,
  ...
)

Arguments

X

A matrix with features as rows and columns as samples.

power

An integer representing the soft-thresholding power to be used to define modules. See the assess_soft_threshold function for aid in selecting this parameter.

corType

The type of correlation to be used to generate a correlation matrix during network formation. One of "pearson" (cor) and "bicor" (bicor).

networkType

The type of network to be generated, either "unsigned", "signed" or "signed hybrid". See blockwiseModules for more details.

module_labels

Specifies whether the modules should be named based on "numbers" or "colours. If module_labels is set to "numbers", then "module_0" represents unclustered genes, whereas if it is set to "colours" then "grey" represents unclustered genes.

maxBlockSize

The chunk size (in terms of the number of features/genes) to process the data. See blockwiseModules for more details. The default (30000) should process standard transcriptomic datasets in a single chunk. Results may differ if the number of features exceeds the chunk size. Lower values of this parameter may use less memory to calculate networks.

verbose

The verbosity, passed to blockwiseModules.

standardise_reduced

If TRUE, the reduced data (eigengenes) are standardised to have a mean of 0 and a standard deviation of 1.

...

Additional arguments to be passed to blockwiseModules.

Value

Returns a list containing:

"E"

The reduced data (eigengenes).

"L"

The module loadings. This represents the values of the PCA rotation matrix for the first principal component generated for each module.

"assignments"

A named vector representing the assignments of genes to modules.

Details

Note that if module_labels is set to "numbers", then "module_0" represents unclustered genes, whereas if it is set to "colours" then "grey" represents unclustered genes.

The function also stores the loadings matrices generated when PCA is performed for each module to calculate eigengenes. These loadings can be used to recalculate the reduced data matrix (eigengenes).

See also

Author

Jack Gisby

Examples

# Get the airway data as a SummarizedExperiment (with a subset of features)
set.seed(2)
airway_se <- ReducedExperiment:::.get_airway_data(n_features = 500)

# Choose an appropriate soft-thresholding power
WGCNA::disableWGCNAThreads()
fit_indices <- assess_soft_threshold(airway_se)
#>    Power SFT.R.sq   slope truncated.R.sq mean.k. median.k. max.k.
#> 1      1   0.9740  4.3700         0.9770   352.0     369.0  408.0
#> 2      2   0.9670  1.9700         0.9690   268.0     286.0  348.0
#> 3      3   0.9540  1.1200         0.9690   213.0     228.0  304.0
#> 4      4   0.9210  0.6790         0.9630   175.0     188.0  270.0
#> 5      5   0.6400  0.3900         0.6710   147.0     156.0  244.0
#> 6      6   0.5540  0.2090         0.6850   126.0     132.0  223.0
#> 7      7   0.0498  0.0600        -0.0944   109.0     113.0  205.0
#> 8      8   0.0348 -0.0448        -0.1660    95.8      97.7  189.0
#> 9      9   0.2320 -0.1320         0.0503    84.9      84.6  176.0
#> 10    10   0.4040 -0.2010         0.2430    75.8      73.7  165.0
#> 11    11   0.5010 -0.2730         0.3590    68.2      64.8  155.0
#> 12    12   0.6120 -0.3300         0.5010    61.7      57.0  146.0
#> 13    13   0.6910 -0.3620         0.6040    56.2      50.6  138.0
#> 14    14   0.8100 -0.4130         0.7560    51.4      45.2  130.0
#> 15    15   0.7550 -0.4760         0.6910    47.1      40.6  124.0
#> 16    16   0.7960 -0.5050         0.7450    43.4      36.4  118.0
#> 17    17   0.7930 -0.5310         0.7470    40.2      32.7  112.0
#> 18    18   0.7640 -0.5600         0.7170    37.3      29.5  107.0
#> 19    19   0.7940 -0.5790         0.7620    34.7      26.8  103.0
#> 20    20   0.8250 -0.5990         0.8040    32.3      24.3   98.4
#> 21    21   0.8500 -0.6250         0.8300    30.2      22.2   94.4
#> 22    22   0.8720 -0.6610         0.8600    28.3      20.3   90.7
#> 23    23   0.8700 -0.6850         0.8550    26.6      18.6   87.2
#> 24    24   0.8700 -0.7000         0.8570    25.0      17.0   84.0
#> 25    25   0.8670 -0.7270         0.8520    23.6      15.6   81.0
#> 26    26   0.8740 -0.7410         0.8690    22.3      14.5   78.1
#> 27    27   0.8620 -0.7590         0.8560    21.1      13.4   75.4
#> 28    28   0.8740 -0.7700         0.8720    19.9      12.4   72.9
#> 29    29   0.8830 -0.7850         0.8770    18.9      11.5   70.5
#> 30    30   0.9040 -0.7980         0.9030    18.0      10.7   68.3
estimated_power <- fit_indices$Power[fit_indices$estimated_power]

# Identify modules using the airway expression matrix
wgcna_res <- run_wgcna(
    assay(airway_se, "normal"),
    verbose = 0,
    power = estimated_power
)

# We find just one module for this small dataset (module_0 indicates unclustered genes)
table(names(wgcna_res$assignments))
#> 
#> module_0 module_1 module_2 module_3 module_4 module_5 
#>       40      265       92       43       35       25