Run WGCNA for a data matrix

Runs WGCNA. Largely a wrapper for the blockwiseModules function. Additionally applies pickSoftThreshold to aid in the selection of the soft thresholding power, reformats data into a format convenient for creating a ModularExperiment object and changes module names from colours to numbers (default).

run_wgcna(
  X,
  powers = 1:30,
  min_r_squared = 0.85,
  max_mean_connectivity = 100,
  corType = "pearson",
  networkType = "signed",
  module_labels = "numbers",
  maxBlockSize = 30000,
  verbose = 0,
  return_full_output = FALSE,
  scale_reduced = TRUE,
  ...
)

Arguments

X

A matrix with features as rows and columns as samples.

powers

The soft-thresholding power(s) to test, see pickSoftThreshold for more details.

min_r_squared

The correlation threshold for selection of the soft-thresholding power, see pickSoftThreshold for more details.

max_mean_connectivity

A threshold for the maximal mean connectivity, used when selecting the soft-thresholding power, see pickSoftThreshold for more details.

corType

The type of correlation to be used to generate a correlation matrix during network formation. One of "pearson" (cor) and "bicor" (bicor).

networkType

The type of network to be generated, either "unsigned", "signed" or "signed hybrid". See blockwiseModules for more details.

module_labels

Specifies whether the modules should be named based on "numbers" or "colours. If module_labels is set to "numbers", then "module_0" represents unclustered genes, whereas if it is set to "colours" then "grey" represents unclustered genes.

maxBlockSize

The chunk size (in terms of the number of features/genes) to process the data. See blockwiseModules for more details. The default (30000) should process standard transcriptomic datasets in a single chunk. Results may differ if the number of features exceeds the chunk size. Lower values of this parameter may use less memory to calculate networks.

verbose

The verbosity, passed to blockwiseModules.

return_full_output

If FALSE (default), returns the results specified below in returns. Else, returns additional information, including "run_wgcna_output") the results specified below in returns; "original_E") the original eigengene matrix calculated by blockwiseModules; "blockwise_modules_output") the object returned by blockwiseModules; and "pick_soft_threshold_output") the output of pickSoftThreshold.

scale_reduced

If TRUE, the reduced data (eigengenes) are standardised to have a mean of 0 and a standard deviation of 1.

...

Additional arguments to be passed to blockwiseModules.

Value

Returns a list containing:

"E"

The reduced data (eigengenes).

"L"

The module loadings. This represents the values of the PCA rotation matrix for the first principal component generated for each module.

"assignments"

A named vector representing the assignments of genes to modules.

Details

The function also stores the loadings matrices generated when PCA is performed for each module to calculate eigengenes. These loadings can be used to quantify the alignment of genes with the module eigengene, can be used to recalculate the reduced data matrix (eigengenes) and these loadings can be used to transform new datasets.

Note that if module_labels is set to "numbers", then "module_0" represents unclustered genes, whereas if it is set to "colours" then "grey" represents unclustered genes.

Note that if powers is a range, the power to use will be selected automatically based on the output of pickSoftThreshold. We recommend that users avoid relying on the automatic threshold detection and instead consider the best soft thresholding power carefully, for instance using the plots described in the WGCNA tutorials.

See also

Examples

# Get the airway data as a SummarizedExperiment (with a subset of features)
set.seed(2)
airway_se <- ReducedExperiment:::.get_airway_data(n_features = 500)

# Identify modules using the airway expression matrix
WGCNA::disableWGCNAThreads()
wgcna_res <- run_wgcna(assay(airway_se, "normal"), verbose = 0, powers = 5)
#>   Power SFT.R.sq slope truncated.R.sq mean.k. median.k. max.k.
#> 1     5   0.0426  -388          -0.22    6.46      6.45   6.52

# We find three modules (module_0 indicates unclustered genes)
table(names(wgcna_res$assignments))
#> 
#> module_0 module_1 module_2 module_3 
#>       42      293      127       38